issue 111 change timing that check file delete grant
git-svn-id: http://xe-core.googlecode.com/svn/branches/1.4.5@8770 201d5d3c-b55e-5fd7-737f-ddc643e51545

+++ modules/file/file.controller.php
@@ -266,7 +266,32 @@
             if(!$_SESSION['upload_info'][$editor_sequence]->enabled) exit();
 
             $upload_target_srl = $_SESSION['upload_info'][$editor_sequence]->upload_target_srl;
-            if($upload_target_srl && $file_srl) $output = $this->deleteFile($file_srl);
+
+			$logged_info = Context::get('logged_info');
+			$oFileModel = &getModel('file');
+
+            $srls = explode(',',$file_srl);
+            if(!count($srls)) return;
+
+            for($i=0;$i<count($srls);$i++) {
+                $srl = (int)$srls[$i];
+                if(!$srl) continue;
+
+                $args = null;
+                $args->file_srl = $srl;
+                $output = executeQuery('file.getFile', $args);
+                if(!$output->toBool()) continue;
+
+                $file_info = $output->data;
+                if(!$file_info) continue;
+
+				$file_grant = $oFileModel->getFileGrant($file_info, $logged_info); 
+
+				if(!$file_grant->is_deletable) continue;
+
+				if($upload_target_srl && $file_srl) $output = $this->deleteFile($file_srl);
+            }
+
         }
 
         /**
@@ -499,8 +524,6 @@
             $srls = explode(',',$file_srl);
             if(!count($srls)) return;
 
-			$oFileModel = &getModel('file');
-			$logged_info = Context::get('logged_info');
             for($i=0;$i<count($srls);$i++) {
                 $srl = (int)$srls[$i];
                 if(!$srl) continue;
@@ -512,10 +535,6 @@
 
                 $file_info = $output->data;
                 if(!$file_info) continue;
-
-				$file_grant = $oFileModel->getFileGrant($file_info, $logged_info); 
-
-				if(!$file_grant->is_deletable) continue;
 
                 $source_filename = $output->data->source_filename;
                 $uploaded_filename = $output->data->uploaded_filename;
